refactor: rename manifest types file
