+++ src/manifest-types.ts
